Python 3 Standard Library: 5.3.10.5 Sizes - Beta, Gamma and Weibull distributions
The Beta distribution is commonly used in Bayesian statistics and applications such as task duration modeling.

The Gamma distribution produced by gammavariate() is used for modeling the sizes of things such as waiting times, rainfall, and computational errors.

The Weibull distribution computed by weibullvariate() is used in failure analysis, industrial engineering, and weather forecasting. It describes the distribution of sizes of particles or other discrete objects.